:children_crossing:(ozi.new): move ``pre/postprocess_arguments`` to ``ozi.new.validate``
