+++ ozi/new/validate.py
@@ -6,6 +6,8 @@
 from __future__ import annotations
 
 import re
+from pathlib import Path
+from typing import TYPE_CHECKING
 from typing import Any
 from typing import Sequence
 from urllib.parse import urlparse
@@ -23,6 +25,13 @@ from ozi.vendor.email_validator import EmailNotValidError
 from ozi.vendor.email_validator import EmailSyntaxError
 from ozi.vendor.email_validator import ValidatedEmail
 from ozi.vendor.email_validator import validate_email
+
+if TYPE_CHECKING:
+    from argparse import Namespace
+    from typing import Callable
+    from typing import TypeAlias
+
+    Composable: TypeAlias = Callable[[Namespace], Namespace]
 
 _CLASSIFIERS = {i.partition(' :: ')[2].strip() for i in classifiers}
 
@@ -231,3 +240,72 @@ def valid_emails(
             case None:  # pragma: no cover
                 continue
     return _author_email, _maintainer_email
+
+
+def _valid_project(project: Namespace) -> Namespace:
+    """Validate a project namespace."""
+    valid_project_name(name=project.name)
+    valid_summary(project.summary)
+    project.license = valid_license(
+        _license=project.license,
+        license_expression=project.license_expression,
+    )
+    valid_home_page(home_page=project.home_page)
+    valid_project_url(project_url=project.project_url)
+    project.copyright_head = valid_copyright_head(
+        copyright_head=project.copyright_head,
+        project_name=project.name,
+        license_file=project.license_file,
+    )
+    valid_spdx(project.license_expression)
+    valid_contact_info(
+        author=project.author,
+        maintainer=project.maintainer,
+        author_email=project.author_email,
+        maintainer_email=project.maintainer_email,
+    )
+    for i in [
+        project.audience,
+        project.environment,
+        project.framework,
+        project.topic,
+    ]:
+        for classifier in i:
+            valid_classifier(classifier)
+    return project
+
+
+def preprocess_arguments(project: Namespace) -> Namespace:
+    """Preprocess (validate) arguments for project namespace."""
+    if project.strict:
+        with TAP.strict():  # pragma: no cover
+            return _valid_project(project)
+    else:
+        return _valid_project(project)
+
+
+def postprocess_arguments(project: Namespace) -> Namespace:
+    """Postprocess (normalize) arguments for project namespace."""
+    project.author_email, project.maintainer_email = valid_emails(
+        author_email=project.author_email,
+        maintainer_email=project.maintainer_email,
+        verify=project.verify_email,
+    )
+    project.keywords = project.keywords.split(',')
+    project.name = re.sub(r'[-_.]+', '-', project.name)
+    project.target = Path(project.target)
+    project.topic = list(set(project.topic))
+    project.dist_requires = list(set(project.dist_requires))
+    if any(
+        i for i in project.target.iterdir() if i not in project.allow_file
+    ):  # defer to good-issue
+        TAP.not_ok('target directory not empty', 'no files will be created', skip=True)
+    match project.ci_provider:
+        case 'github':  # pragma: no cover
+            pass
+        case _:
+            TAP.not_ok(
+                f'--ci-provider "{project.ci_provider}" unrecognized. ci_user will not be set.',
+            )
+    project.allow_file = set(map(Path, project.allow_file))
+    return project
